+++ comb_spec_searcher/exception.py
@@ -1,0 +1,6 @@
+class InsaneTreeError(Exception):
+    pass
+
+
+class TaylorExpansionError(Exception):
+    pass